fix(android): refactor accessibility
